FEAT: add Open API security definitions as project option
A new project option is introduced to include the security responses to
authorized API endpoints.
This feature required to have an additional controller to be tested
correctly, which means the `WebApiProject` was slightly altered to have
the possibility to add fixture files after the project is created but
before it's started.